+++ fess-crawler/src/main/java/org/codelibs/fess/crawler/client/smb/SmbClient.java
@@ -152,7 +152,7 @@ public class SmbClient extends AbstractCrawlerClient {
         TimeoutTask accessTimeoutTask = null;
         if (accessTimeout != null) {
             accessTimeoutTarget = new AccessTimeoutTarget(Thread.currentThread());
-            accessTimeoutTask = TimeoutManager.getInstance().addTimeoutTarget(accessTimeoutTarget, accessTimeout.intValue(), false);
+            accessTimeoutTask = TimeoutManager.getInstance().addTimeoutTarget(accessTimeoutTarget, accessTimeout, false);
         }
 
         try {
