+++ fess-crawler/src/main/java/org/codelibs/fess/crawler/extractor/impl/ApiExtractor.java
@@ -195,7 +195,7 @@ public class ApiExtractor extends AbstractExtractor {
         TimeoutTask accessTimeoutTask = null;
         if (accessTimeout != null) {
             accessTimeoutTarget = new AccessTimeoutTarget(Thread.currentThread());
-            accessTimeoutTask = TimeoutManager.getInstance().addTimeoutTarget(accessTimeoutTarget, accessTimeout.intValue(), false);
+            accessTimeoutTask = TimeoutManager.getInstance().addTimeoutTarget(accessTimeoutTarget, accessTimeout, false);
         }
 
         final ExtractData data = new ExtractData();
